refactor: :recycle: decompose patient registration form into subcomponents
